Add ColorHeatMap class to ES-GUI project
A new class, 'ColorHeatMap', is introduced to the ES-GUI project. This class will be responsible for creating heat maps based on specified color values. It includes methods for initialising color blocks from different sources (path, array, default), and a method for getting a color value based on a double value within a provided range.